+++ FFXIVAPP.Plugin.Parse/Models/LinkedStats/MinStat.cs
@@ -34,17 +34,17 @@ namespace FFXIVAPP.Plugin.Parse.Models.LinkedStats
 {
     public sealed class MinStat : LinkedStat
     {
-        public MinStat(string name, params Stat<decimal>[] dependencies) : base(name, 0m)
+        public MinStat(string name, params Stat<double>[] dependencies) : base(name, 0)
         {
             AddDependency(dependencies[0]);
             GotValue = false;
         }
 
-        public MinStat(string name, decimal value) : base(name, 0m)
+        public MinStat(string name, double value) : base(name, 0)
         {
         }
 
-        public MinStat(string name) : base(name, 0m)
+        public MinStat(string name) : base(name, 0)
         {
         }
 
@@ -57,8 +57,8 @@ namespace FFXIVAPP.Plugin.Parse.Models.LinkedStats
         /// <param name="newValue"> </param>
         public override void DoDependencyValueChanged(object sender, object previousValue, object newValue)
         {
-            var ovalue = (decimal) previousValue;
-            var nvalue = (decimal) newValue;
+            var ovalue = (double) previousValue;
+            var nvalue = (double) newValue;
             var delta = Math.Max(ovalue, nvalue) - Math.Min(ovalue, nvalue);
             if ((delta >= Value) && GotValue)
             {
