+++ FFXIVAPP.Plugin.Parse/Models/StatGroups/Player.Stats.DamageOverTime.cs
@@ -50,7 +50,7 @@ namespace FFXIVAPP.Plugin.Parse.Models.StatGroups
                 return;
             }
 
-            var currentDamage = line.Crit ? line.Amount > 0 ? ParseHelper.GetOriginalAmount(line.Amount, (decimal) .5) : 0 : line.Amount;
+            var currentDamage = line.Crit ? line.Amount > 0 ? ParseHelper.GetOriginalAmount(line.Amount, (double) .5) : 0 : line.Amount;
             if (currentDamage > 0)
             {
                 ParseHelper.LastAmountByAction.EnsurePlayerAction(line.Source, line.Action, currentDamage);
